Update shebang and `mktemp` usage  to work on both Linux and Mac
Annoyingly the GNU mktemp and BSD mktemp are not compatable in usage.
This formualation works to the create the temp folder under $TMPDIR,
but GNU interpretes `faust2er301.XXXXX` as a template and replaces the
`XXXXX` with some hash, whereas BSD interprets it as a prefix and
adds a hash on the end. Slightly wierd, but it works in both cases.